backend/requirements.txt
railtracks
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
websockets==13.1
httpx==0.27.2
openai==1.51.2
python-dotenv==1.0.1
ibm-watsonx-ai>=1.4.0
pydantic>=2.9.2
supabase
resend